+++ src/rpc/dataacq.cpp
@@ -4,16 +4,16 @@
 // Distributed under the MIT/X11 software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include "main.h"
-#include "rpcserver.h"
-#include "kernel.h"
-#include "block.h"
-#include "txdb.h"
-#include "appcache.h"
-#include "neuralnet/claim.h"
-#include "neuralnet/quorum.h"
-#include "neuralnet/superblock.h"
-#include "util.h"
+#include "../main.h"
+#include "server.h"
+#include "../kernel.h"
+#include "../block.h"
+#include "../txdb.h"
+#include "../appcache.h"
+#include "../neuralnet/claim.h"
+#include "../neuralnet/quorum.h"
+#include "../neuralnet/superblock.h"
+#include "../util.h"
 
 #include <boost/filesystem.hpp>
 #include <iostream>
