+++ src/rpc/dump.cpp
@@ -5,11 +5,11 @@
 #include <iostream>
 #include <fstream>
 
-#include "init.h" // for pwalletMain
-#include "rpcserver.h"
-#include "rpcprotocol.h"
-#include "ui_interface.h"
-#include "base58.h"
+#include "../init.h" // for pwalletMain
+#include "server.h"
+#include "protocol.h"
+#include "../ui_interface.h"
+#include "../base58.h"
 
 #include <boost/date_time/posix_time/posix_time.hpp>
 #include <boost/variant/get.hpp>
