+++ src/rpc/misc.cpp
@@ -3,8 +3,8 @@
 // Distributed under the MIT/X11 software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include "rpcprotocol.h"
-#include "util.h"
+#include "protocol.h"
+#include "../util.h"
 
 #include <univalue.h>
