+++ src/rpc/net.cpp
@@ -2,15 +2,15 @@
 // Distributed under the MIT/X11 software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include "rpcserver.h"
-#include "rpcprotocol.h"
-#include "alert.h"
-#include "wallet.h"
-#include "db.h"
-#include "streams.h"
-#include "walletdb.h"
-#include "net.h"
-#include "banman.h"
+#include "server.h"
+#include "protocol.h"
+#include "../alert.h"
+#include "../wallet.h"
+#include "../db.h"
+#include "../streams.h"
+#include "../walletdb.h"
+#include "../net.h"
+#include "../banman.h"
 
 using namespace std;
