+++ src/scraper/scraper.h
@@ -23,7 +23,7 @@
 
 #include <memory>
 #include "net.h"
-#include "rpcprotocol.h"
+#include "rpc/protocol.h"
 
 // See fwd.h for certain forward declarations that need to be included in other areas.
 #include "fwd.h"
