+++ src/wallet.cpp
@@ -14,9 +14,9 @@
 #include <boost/algorithm/string/replace.hpp>
 #include <boost/thread.hpp>
 #include "block.h"
-#include "rpcserver.h"
-#include "rpcclient.h"
-#include "rpcprotocol.h"
+#include "rpc/server.h"
+#include "rpc/client.h"
+#include "rpc/protocol.h"
 #include <boost/variant/apply_visitor.hpp>
 #include <script.h>
 #include "main.h"
